- Adding tests for AgentRepository.buildSql - Cleaning up the repo layer to get rid of unnecessary abstraction (at this point) - Fix a bug in TaskAllocationDetailRepository.delete method

+++ src/main/java/ffam/task/data/TaskAllocationDetailRepository.java
@@ -1,11 +1,11 @@
 package ffam.task.data;
 
-import ffam.general.Sql;
 import ffam.task.domain.TaskAllocationDetail;
 import ffam.task.domain.TaskPriority;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
@@ -18,25 +18,18 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static ffam.general.StatementAttribute.Type.API;
-import static ffam.general.StatementAttribute.Type.DATA;
-import static ffam.general.StatementAttribute.attribute;
-import static ffam.general.StatementInfo.withAttributes;
-
 @Slf4j
 @Repository
 public class TaskAllocationDetailRepository {
-    private final Sql sql;
+    private final JdbcTemplate jdbcTemplate;
 
     @Autowired
-    public TaskAllocationDetailRepository(Sql sql) {
-        this.sql = sql;
+    public TaskAllocationDetailRepository(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
     public List<TaskAllocationDetail> findAll() {
-        List<TaskAllocationDetail> results = sql.query(
-                withAttributes(attribute(API, "TaskAllocationDetailRepository.findAll")
-                ),
+        List<TaskAllocationDetail> results = jdbcTemplate.query(
                 "SELECT AGENT_ID, TASK_ID, TASK_PRIORITY, DATE_CREATED " +
                         "FROM TASK_ALLOCATION " +
                         "ORDER BY DATE_CREATED DESC ",
@@ -47,10 +40,7 @@ public class TaskAllocationDetailRepository {
     }
 
     public Optional<TaskAllocationDetail> findByAgentId(String agentId) {
-        List<TaskAllocationDetail> results = sql.query(
-                withAttributes(attribute(API, "TaskAllocationDetailRepository.findByAgentId"),
-                        attribute(DATA, agentId)
-                ),
+        List<TaskAllocationDetail> results = jdbcTemplate.query(
                 "SELECT AGENT_ID, TASK_ID, TASK_PRIORITY, DATE_CREATED " +
                         "FROM TASK_ALLOCATION WHERE AGENT_ID = ? ",
                 new Object[]{agentId},
@@ -66,12 +56,7 @@ public class TaskAllocationDetailRepository {
         val taskAllocationDetailOptional = findByAgentId(agentId);
 
         if (!taskAllocationDetailOptional.isPresent()) {
-            int count = sql.update(
-                    withAttributes(
-                            attribute(API, "TaskAllocationDetailRepository.createOrUpdate"),
-                            attribute(DATA, agentId),
-                            attribute(DATA, taskId)
-                    ),
+            int count = jdbcTemplate.update(
                     "INSERT INTO TASK_ALLOCATION (TASK_ID, AGENT_ID, TASK_PRIORITY) VALUES (?, ?, ?)",
                     new Object[]{taskId, agentId, taskPriority == TaskPriority.HIGH ? 1 : 0}
             );
@@ -81,12 +66,7 @@ public class TaskAllocationDetailRepository {
             List<String> taskIdList = taskAllocationDetailOptional.get().getTaskId();
             taskIdList.add(taskId);
 
-            int count = sql.update(
-                    withAttributes(
-                            attribute(API, "TaskAllocationDetailRepository.createOrUpdate"),
-                            attribute(DATA, agentId),
-                            attribute(DATA, taskId)
-                    ),
+            int count = jdbcTemplate.update(
                     "UPDATE TASK_ALLOCATION " +
                             "SET TASK_ID = ? ," +
                             "TASK_PRIORITY = ? , " +
@@ -116,36 +96,32 @@ public class TaskAllocationDetailRepository {
             List<String> taskIdList = taskAllocationDetailOptional.get().getTaskId();
             taskIdList.remove(taskId);
 
-            int count = sql.update(
-                    withAttributes(
-                            attribute(API, "TaskAllocationDetailRepository.createOrUpdate"),
-                            attribute(DATA, agentId),
-                            attribute(DATA, taskId)
-                    ),
-                    "UPDATE TASK_ALLOCATION " +
-                            "SET TASK_ID = ? ," +
-                            "TASK_PRIORITY = ? , " +
-                            "DATE_CREATED = ? " +
-                            "WHERE AGENT_ID = ? ",
-                    new Object[]{String.join(",", taskIdList),
-                            taskAllocationDetail.getTaskPriority() == TaskPriority.HIGH ? 1 : 0, // If already present and updating, take newer state
-                            Date.from(ZonedDateTime.now().toInstant()),
-                            agentId}
-            );
-
-            return count > 0;
-        } else {
-            int count = sql.update(
-                    withAttributes(
-                            attribute(API, "TaskAllocationDetailRepository.delete"),
-                            attribute(DATA, taskId)
-                    ),
-                    "DELETE FROM TASK_ALLOCATION WHERE TASK_ID = ? ",
-                    new Object[]{taskId}
-            );
-            return count > 0;
+            if(taskIdList.isEmpty()){
+                // No more tasks left for the agent. Clean the record up
+                int count = jdbcTemplate.update(
+                        "DELETE FROM TASK_ALLOCATION WHERE TASK_ID = ? ",
+                        new Object[]{taskId}
+                );
+                return count > 0;
+            }else {
+                // More tasks left for the agent. Keep the record here.
+                int count = jdbcTemplate.update(
+                        "UPDATE TASK_ALLOCATION " +
+                                "SET TASK_ID = ? ," +
+                                "TASK_PRIORITY = ? , " +
+                                "DATE_CREATED = ? " +
+                                "WHERE AGENT_ID = ? ",
+                        new Object[]{String.join(",", taskIdList),
+                                taskAllocationDetail.getTaskPriority() == TaskPriority.HIGH ? 1 : 0, // If already present and updating, take newer state
+                                Date.from(ZonedDateTime.now().toInstant()),
+                                agentId}
+                );
+                return count > 0;
+            }
         }
 
+        //Not A Task for this agent
+        return false;
     }
 
     //region private
